requirements.txt
python-docx
bs4
requests
transformers
pandas
numpy
scikit-learn
nltk
beautifulsoup4
openpyxl
flask
sentence-transformers
